Push FOLIO component previews to a namespace.
