+++ spec/component_previews/folio/request_component_preview.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Folio
+  class RequestComponentPreview < ViewComponent::Preview
+    layout 'lookbook'
+
+    def default; end
+  end
+end
