--- spec/component_previews/request_component_preview.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class RequestComponentPreview < ViewComponent::Preview
-  layout 'lookbook'
-
-  def default; end
-end
